requirements.txt
Flask
flask-cors
flask-sslify
requests
python-dotenv
weasyprint
selenium
webdriver_manager